// src/types/processed/MurderMystery.ts
export interface MurderMystery {
    activeCosmetics: {
        animatedHat: string
        deathCry: string
        gesture: string
        killNote: string
        knifeSkin: string
        lastWords: string
        projectileTrail: string
        victoryDance: string
    }
    alphaChance: number
    bowKills: number
    chestHistoryNew: string[]
    coins: number
    coinsPickedUp: number
    deaths: number
    detectiveChance: number
    detectiveWins: number
    gamesPlayed: number
    grantedChests: number
    kills: number
    killsAsInfected: number
    killsAsMurderer: number
    killsAsSurvivor: number
    knifeKills: number
    timesLastAlive: number
    chestHistory: string[]
    chests: number
    murdererChance: number
    murdererWins: number
    books: string[]
    openedChests: number
    packages: string[]
    quickestDetectiveWinTime: number
    quickestMurdererWinTime: number
    quickestShowdownWinTime: number
    shopSort: string
    showdownPotg: number
    showQueueBook: boolean
    suicides: number
    survivorWins: number
    thrownKnifeKills: number
    totalTimeSurvived: number
    trapKills: number
    timesWasHero: number
    wins: number
}